refactor(web): split server.rs by responsibility
big-code-analysis-web/src/web/server.rs held the actix entry point, all
~40 route handlers, the error types, and the route-registration / guard
/ fallback machinery in one 1677-SLOC file, over the 800-line self-scan
limit. Split it by responsibility into a server/ subdirectory:

  - server/handlers.rs the endpoint handlers + request/response structs
  - server/errors.rs   ParseError/BodyError + ResponseError impls,
                       error_kind, the message constants, error helpers
  - server/routing.rs  register_endpoints, configure_routes, the content
                       guards, the method/not-found fallbacks, ROUTES

Pure code movement: bodies verbatim, no logic, route, status-code, or
content-negotiation change. server.rs keeps the entry points (`run`,
`run_with_timeout`), `run_parse`, `ParseConfig`, the core size/timeout
constants, and the inline test module (its `#[path]` include is
unchanged). Each submodule uses `use super::*` and server.rs re-exports
them so the route registration and tests reach every handler/error by
its bare name. Visibility was widened to pub(crate) only for items
reached across the new boundary (handlers wired by register_endpoints,
error helpers, configure_routes). The declaring module (web/mod.rs) is
byte-identical.

server.rs 1677 -> 309 production SLOC; handlers.rs 633, errors.rs 463,
routing.rs 306.
